change stuctor project
